Move remaining plugins
