+++ setup.py
@@ -130,7 +130,7 @@ setup(
     version=_get_version(),
     cmdclass=cmdclass,
     ext_modules=[
-        Extension("builtin_plugins.plugin.amulet_team_3d_viewer._view_3d._view_3d", [])
+        Extension("builtin_plugins.plugin.amulet.view_3d._view_3d._view_3d", [])
     ]
     * (not os.environ.get("AMULET_SKIP_COMPILE", None)),
     install_requires=requirements.get_runtime_dependencies(),
