+++ src/builtin_plugins/plugin/amulet/resource_pack/__init__.py
