+++ src/builtin_plugins/plugin/amulet/resource_pack/_api.py
