+++ src/builtin_plugins/plugin/amulet/resource_pack/plugin.json
@@ -1,5 +1,6 @@
 {
-	"identifier": "amulet_team_resource_pack",
+	"namespace": "amulet",
+	"identifier": "resource_pack",
 	"version": "1.0.0",
 	"name": "Amulet Resource Pack",
 	"depends": {
