+++ src/builtin_plugins/plugin/amulet/resource_pack/resources/lang/en.lang
