+++ src/builtin_plugins/plugin/amulet/view_3d/_view_3d/_chunk_mesher_lod0.py.cpp
