+++ src/builtin_plugins/plugin/amulet/view_3d/_view_3d/_chunk_mesher_lod0.pyi
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
 import amulet.core.chunk.component.block_component
-import plugin.amulet_team_3d_viewer._view_3d._resource_pack_base
+import plugin.amulet.view_3d._view_3d._resource_pack_base
 
 __all__ = ["create_lod0_chunk"]
 
 def create_lod0_chunk(
-    resource_pack: plugin.amulet_team_3d_viewer._view_3d._resource_pack_base.AbstractOpenGLResourcePack,
+    resource_pack: plugin.amulet.view_3d._view_3d._resource_pack_base.AbstractOpenGLResourcePack,
     cx: int,
     cz: int,
     block_component: amulet.core.chunk.component.block_component.BlockComponentData,
