+++ src/builtin_plugins/plugin/amulet/view_3d/_view_3d/_resource_pack_base.pyi
